encore plein de trucs

+++ src/projets/projets.service.ts
@@ -5,12 +5,16 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Projet } from './entities/projet.entity';
 import { Repository } from 'typeorm';
 import { plainToInstance } from 'class-transformer';
+import { User } from 'src/users/entities/user.entity';
+import { RoleEnum } from 'src/users/users.role_enum';
 
 @Injectable()
 export class ProjetsService {
   constructor(
     @InjectRepository(Projet)
     private projetsRepository: Repository<Projet>,
+    @InjectRepository(User)
+    private usersRepository: Repository<User>,
   ) {}
 
   async create(createProjetDto: CreateProjetDto) {
@@ -33,10 +37,19 @@ export class ProjetsService {
   }
 
   update(id: string, updateProjetDto: UpdateProjetDto) {
-    return `This action updates a #${id} projet`;
+    return this.projetsRepository.update(id, updateProjetDto).then(() => {
+      return this.findOne(id);
+    });
   }
 
-  async remove(id: string) {
+  async remove(id: string, userId: string) {
+    const user = await this.usersRepository.findOne({ where: { id: userId } });
+    const projet = await this.projetsRepository.findOne({ where: { id } });
+    
+    if (projet?.ownerId !== userId && user?.role !== RoleEnum.ADMIN) {
+      throw new Error('Vous n\'avez pas l\'autorisation de supprimer ce projet.');
+    }
+
     await this.projetsRepository.delete(id);
   }
 }
